Create README.md
the project can be used to experiment or as a template for a new project
Single html pages applications offer the benefit of loading the whole app in one go providing a better user experience. The end-user gets the illusion of seeing a website/app with different views.
the app is also responsive and resizes according to CSS media queries. 
I have used some different background colors to highlight  the changes in sizes.
Those colors can be changed easily in the css part of the project styles.css.

the main features showcased in the project are:

+unique  loading
+responsiveness of the app according to different devices sizes
+swapping of the navigation bar : on top of screen for desktop devices  and at the bottom for handheld devices
+multiple columns divs for desktop devices to 2 column or 1 column divs for mobile devices.

I started working on responsiveness and media queries after reading the work of Ethan Marcotte.
I have used the techniques on many occasions including: 

http://isabellebidou.com/mobile.html 
http://isabellebidou.com/dance/